+++ packtools/__init__.py
@@ -1,5 +1,8 @@
+"""Packtools is a Python library and set of command line utilities which can be
+used to handle SciELO Publishing Schema packages and XML files.
+"""
 from .xray import SPSPackage
-from .stylechecker import XMLValidator
+from .domain import XMLValidator, XMLPacker
 from .utils import XML
-from .packbuilder import XMLPacker
+from .version import __version__