+++ tests/test_xmlvalidator.py
@@ -5,8 +5,7 @@ from tempfile import NamedTemporaryFile
 
 from lxml import etree, isoschematron
 
-from packtools import stylechecker
-from packtools.adapters import StyleError
+from packtools import domain, style_errors
 
 
 # valid: <a><b></b></a>
@@ -52,17 +51,17 @@ class XMLValidatorTests(unittest.TestCase):
 
     @setup_tmpfile
     def test_initializes_with_filepath(self):
-        self.assertTrue(stylechecker.XMLValidator(self.valid_tmpfile.name, no_doctype=True))
+        self.assertTrue(domain.XMLValidator(self.valid_tmpfile.name, no_doctype=True))
 
     def test_initializes_with_etree(self):
         fp = io.BytesIO(b'<a><b>bar</b></a>')
         et = etree.parse(fp)
 
-        self.assertTrue(stylechecker.XMLValidator(et, no_doctype=True))
+        self.assertTrue(domain.XMLValidator(et, no_doctype=True))
 
     def test_validation(self):
         fp = etree.parse(io.BytesIO(b'<a><b>bar</b></a>'))
-        xml = stylechecker.XMLValidator(fp, no_doctype=True)
+        xml = domain.XMLValidator(fp, no_doctype=True)
         xml.dtd = etree.XMLSchema(etree.parse(sample_xsd))
 
         result, errors = xml.validate()
@@ -71,7 +70,7 @@ class XMLValidatorTests(unittest.TestCase):
 
     def test_invalid(self):
         fp = etree.parse(io.BytesIO(b'<a><c>bar</c></a>'))
-        xml = stylechecker.XMLValidator(fp, no_doctype=True)
+        xml = domain.XMLValidator(fp, no_doctype=True)
         xml.dtd = etree.XMLSchema(etree.parse(sample_xsd))
 
         result, _ = xml.validate()
@@ -80,16 +79,16 @@ class XMLValidatorTests(unittest.TestCase):
     def test_invalid_errors(self):
         # Default lxml error log.
         fp = etree.parse(io.BytesIO(b'<a><c>bar</c></a>'))
-        xml = stylechecker.XMLValidator(fp, no_doctype=True)
+        xml = domain.XMLValidator(fp, no_doctype=True)
         xml.dtd = etree.XMLSchema(etree.parse(sample_xsd))
 
         _, errors = xml.validate()
         for error in errors:
-            self.assertIsInstance(error, StyleError)
+            self.assertIsInstance(error, style_errors.StyleError)
 
     def test_annotate_errors(self):
         fp = etree.parse(io.BytesIO(b'<a><c>bar</c></a>'))
-        xml = stylechecker.XMLValidator(fp, no_doctype=True)
+        xml = domain.XMLValidator(fp, no_doctype=True)
         xml.dtd = etree.XMLSchema(etree.parse(sample_xsd))
 
         err_xml = xml.annotate_errors()
@@ -99,7 +98,7 @@ class XMLValidatorTests(unittest.TestCase):
 
     def test_validation_schematron(self):
         fp = etree.parse(io.BytesIO(b'<Total><Percent>70</Percent><Percent>30</Percent></Total>'))
-        xml = stylechecker.XMLValidator(fp, no_doctype=True)
+        xml = domain.XMLValidator(fp, no_doctype=True)
         xml.schematron = isoschematron.Schematron(etree.parse(sample_sch))
 
         result, errors = xml._validate_sch()
@@ -108,7 +107,7 @@ class XMLValidatorTests(unittest.TestCase):
 
     def test_invalid_schematron(self):
         fp = etree.parse(io.BytesIO(b'<Total><Percent>60</Percent><Percent>30</Percent></Total>'))
-        xml = stylechecker.XMLValidator(fp, no_doctype=True)
+        xml = domain.XMLValidator(fp, no_doctype=True)
         xml.schematron = isoschematron.Schematron(etree.parse(sample_sch))
 
         result, errors = xml._validate_sch()
@@ -117,7 +116,7 @@ class XMLValidatorTests(unittest.TestCase):
 
     def test_annotate_errors_schematron(self):
         fp = etree.parse(io.BytesIO(b'<Total><Percent>60</Percent><Percent>30</Percent></Total>'))
-        xml = stylechecker.XMLValidator(fp, no_doctype=True)
+        xml = domain.XMLValidator(fp, no_doctype=True)
         xml.schematron = isoschematron.Schematron(etree.parse(sample_sch))
         xml.dtd = etree.XMLSchema(etree.parse(sample_xsd))
 
@@ -130,11 +129,11 @@ class XMLValidatorTests(unittest.TestCase):
         fp = io.BytesIO(b'<a><b>bar</b></a>')
         et = etree.parse(fp)
 
-        self.assertRaises(ValueError, lambda: stylechecker.XMLValidator(et, no_doctype=False))
+        self.assertRaises(ValueError, lambda: domain.XMLValidator(et, no_doctype=False))
 
     def test_checks_allowed_doctype_public_ids(self):
         fp = io.BytesIO(b'<?xml version="1.0" encoding="UTF-8"?><!DOCTYPE article PUBLIC "-//NLM//DTD JATS (Z39.96) Journal Publishing DTD v1.0 20120330//EN" "JATS-journalpublishing1.dtd"><a><b>bar</b></a>')
         et = etree.parse(fp)
 
-        self.assertTrue(stylechecker.XMLValidator(et, no_doctype=False))
+        self.assertTrue(domain.XMLValidator(et, no_doctype=False))